LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_space-expected.txt

<<<<<<< HEAD
Blocked access to external URL http://www.localhost:8800/resource-timing/resources/TAOResponse.py?tao=space
CONSOLE MESSAGE: NetworkError:  A network error occurred.
CONSOLE MESSAGE: Error: assert_greater_than: The iframe should have at least one resource timing entry. expected a number greater than 0 but got 0
=======
>>>>>>> 30d2ff3274ef ([COCOA] Add support for web-platform-tests custom domains when running layout tests)
Description

This test validates that for a cross origin resource, the timing allow check algorithm will fail if the Timing-Allow-Origin header value list contains [origin|*]{2,}, separated by space.


PASS window.performance is defined
PASS window.performance.getEntriesByType is defined
PASS redirectStart, redirectEnd, domainLookupStart, domainLookupEnd, connectStart, connectEnd, secureConnectionStart, requestStart, and responseStart -- should be all returned as 0 when the Timing-Allow-Origin header value of the HTTP response is a space separated origin/wildcard list
